refactor: cleanup codebase
