+++ src/data/types/range.ts
@@ -1,7 +1,7 @@
 import { Tagged } from "../../cbor";
 import { SurrealDbError } from "../../errors";
 import { toSurrealqlString } from "../../util/to-surrealql-string";
-import { TAG_BOUND_EXCLUDED, TAG_BOUND_INCLUDED, TAG_RANGE } from "../cbor";
+import { TAG_BOUND_EXCLUDED, TAG_BOUND_INCLUDED } from "../cbor";
 import {
 	type RecordIdValue,
 	escape_id_part,
