+++ src/engines/abstract.ts
@@ -1,12 +1,5 @@
-import type { Encoded } from "../cbor";
 import type { EngineDisconnected } from "../errors";
-import type {
-	LiveAction,
-	LiveHandlerArguments,
-	Patch,
-	RpcRequest,
-	RpcResponse,
-} from "../types";
+import type { LiveHandlerArguments, RpcRequest, RpcResponse } from "../types";
 import type { Emitter } from "../util/emitter";
 
 export type Engine = new (context: EngineContext) => AbstractEngine;
